a64/nbrom: Add first batch of sections and symbols
The A64 NBROM has mostly the same structure as the H5 NBROM, with only a
few functions moved. Obvious functional changes are the handling of a
few additional bits in the LCJS and differences in DMA handling.

Again, functions were first found from bl instructions, and then some
manually after that. Unanalyzed functions are named with their address.
Symbols are suffixed with the driver they belong in to avoid duplicates.