scripts/callgraph: Initial version
This script generates a dot graph description from all of the call
instructions in an assembly listing. It uses the most recently seen
symbol as the caller, and the operand symbol as the callee. If the
operand does not have a symbol, it falls back to a generic function
name.

This script assumes all call instructions (or architecture equivalent)
jump to the beginning of a function.