scripts/gensyms: Stricter address matching
On ARM, indirect branches to the "fp" register erroneously matched the
previous regex, since the leading "f" is a hex digit. Require two hex
digits in a row to avoid this. This requires adding additional logic to
skip "0x", which otherwise only has a single leading hex digit.

To avoid the regex getting unreadable, switch to a POSIX extended regex.