sun55iw3p1: add bare Allwinner A523 SBROM and MBROM dump
This is from a Teclast P85T tablet with that SoC, and the secure fuse
burnt. Reading the BROM from non-secure state (via FEL) returns all
zeroes, but putting some code in a TOC0 image on an SD card gets that
executed in secure state. This allows copying the BROM to SRAM, then
going to FEL, and read the copy from there.

SBROM is located at address 0x0, and is the primary (secure) reset entry
point. Allwinner calls the second blob MBROM ("Medium Boot module"),
It is located at address 0x10000, so in the gap between SBROM and SRAM.

Just the hexdump for now.